Saving a project almost works, but the DnD from tree to playpen CLONES THE ENTIRE SOURCE DATABASE every time you drag and drop!
The biggest problem with this behaviour is that the SourceColumn
property on the playpen tables point into the cloned database rather
than the actual source SQLColumn instances in the DBTree.  That screws
up saving and makes all the source columns show up as "null".

This obviously needs to be fixed, and it hopefully will be fixed shortly.


git-svn-id: http://power-architect.googlecode.com/svn/trunk@95 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87