Big change:
The Architect now uses a different mechanism for loading JDBC drivers and connecting to databases.  There is a new ArchitectDataSourceType class which acts as a kind of parent or supertype of an ArchitectDataSource.  The type contains its own classloader and classpath, and the data source itself (with the help of its parent type) can create its own connections.

There is also a new GUI for creating and modifying data source types, as well as the corresponding modifications to the DBCSPanel.

This change touches a lot of classes, which needed little tweaks to conform to the new business and gui API's.


git-svn-id: http://power-architect.googlecode.com/svn/trunk@1466 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87