Fixed a bug that was causing an error message to pop up if the user cancelled a project file load by pressing the cancel button on the progress dialog.
The ProgressMonitorInputStream throws an InterruptedIOException with the message "progress" when this happens, and the LoadFileWorker didn''t catch this. So in the LoadFileWorker's cleanup, it checks if doStuffException was an InterruptedIOException with the message "progress". If so, it ignores it, assuming that the user cancelled the load. Otherwise, it shows an error.

Also changed some logging messages in SwingUIProject which incorrectly stated that errors were related to the 'config' file, when they were really with the project file


git-svn-id: http://power-architect.googlecode.com/svn/trunk@1570 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87