*** Code Bomb Alert ***
We've moved all the ArchitectDataSource family of classes into the SQL Power Library project.
This project will be going up on Google Code this week.  For now, we're committing the new
binary of sqlpower_library.jar, which includes the transplanted code.

The ArchitectDataSource family of classes has been renamed SPDataSource (for SQL Power)
since it's no longer Architect specific.

By the way, the point here is to pull out code that is not inherently Architect-specific,
and which is actually being used by other SQL Power apps.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@1664 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87