Fix for bug 1346: http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1346
The problem was that neither save action was setting the SwingUIProject.file field when opening a file. 

Also fixed the bug where opening a project from the command-line or by clicking on a project file in the OS gui didn't add that file to the recent menu. 

TODO: Make a unit test to ensure that after loading a file, the file field in SwingUIProject is getting set.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@1747 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87