Fixed bug 1178
Added a right click menu item to the dbtree in the architect frame. Now when right clicking on a schema object in the tree there is an option to "Compare to current" which opens the Compare Data Models Dialog and sets the correct options for comparing it to the contents of the playpen.

The change to the database dropdown menu is done in a method called from the action event, updates to the catalog dropdown menu and the schema dropdown menu are done in the clean up methods of the schema and catalog populators.

Changes:

Moved the CompareDM dialog that that was created each time the button was pressed to its own class, so that its values can be changed by the right click action, and only one instance exists per ArchitectFrame.

Added more checks to the refresh menu method in DBTree so it now adds the "Compare to Current" menu only if a schema is selected, or a catalog is selected and the Database platform does not have schemas, or a database is selected that only has tables.

Added a CompareToCurrentAction class in DBTree that handles the right click action

Added two string values in to the CatalogPopulator and one to SchemaPopulator so that in its cleanup it can set the dropdown boxes to a specified value.

Added a compareCurrentToOrig method in both CompareDMDialog and CompareDMPanel, the one in ComapreDMPanel just passes the parameters on to the one in CompareDMFrame. This method sets the values in the Panel to select the specified objects.

Moved the creation of the CompareDMSettings before the creation of the frame in TestingArchitectSwingSession  because the ComapreDMPanel is now created with the frame and the setting need to be created first.

Added tests in TestDBTree to check that the menu item is being generated at the right time.

Note:

This is the first revision where the default 64mb for the heap is not enough to run the mega test suite, use the -Xmx##m argument in the run dialog where ## is the size in mb that is used to fix this.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@1782 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87