Fixed bug 1321
Changed the make dialog method for the new column dialog so that it can add the column to the table iff the ok button is pressed for the inital creation, it is unaffected in other cases. This was done so that pushing cancel after creating a new column did not leave a "new_Column" column.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@1812 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87