Fixed bugs created from r1835
Released the constraints on auto-increment in foreign keys, although this does not seem to be useful in any situation we could think of, it is included to increase flexibility for the user.

Changed the available combos for column options, to allow for more legal but questionable options, again to increase user flexibility.

Re-enabled the test cases and firing the event for setting auto-increment. The event is ignored for foreign key child column.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@1838 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87