Changed the Compound undo event to not use AWT.event so that kettle will not yell at us. Removed the target object from the event because it was only used to pass to the super constructor.
In the library we changed the SPDataSource to have its own method to look for a kettle name so that we can keep adding them as kettle changes them for no real reason.

Now the names are stored as a : delimited list and are returned as a set.

Added a close method to the profile manager so that it can be stopped and will not lose threads.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@1911 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87