Fixed bug 1425, which is actually a MatchMaker bug, but this fix should benefit everyone who uses MySQL with the Architect.
MySQL identifiers are case-sensitive when the database is on a case-sensitive filesystem (as most Unix filesystems are).  This means you're in for a mess when reverse engineering a database from an all-uppercase platform such as Oracle, or a mixed-case case insensitive platform like SQL Server.  Your MySQL table names will be really hard to get right!

The fix was simply to downcase all identifiers in the DDL generator.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@1941 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87