New refactoring of the profiling API.
The main reason for refactoring is that we want to be able to plug in different profiling implementations.  The current implementation (same as always) relies on the remote database to do most of the heavy lifting. Now it's possible to give the user the option to use a new implementation (not yet written) that gets the local JVM to do all the work, and the database just does a SELECT *.

The main change in this refactoring was to get rid of the populate() method in TableProfileResult.  There is a new interface, TableProfileCreator, which defines a doProfile(tableResult) method that does the work tableResult.populate() used to do.  Hence, the ProfileManager can now choose any implementation of TableProfileCreator it wants to for each profile it creates.

There isn't a user interface for choosing between profiling implementations yet. That's coming soon.  As is a new ProfileCreator implementation.

This commit also comes with some bug fixes:

-The user pref for "N" in the Top "N" values is now honoured
-Added a lot more checks for isCancelled(), so the response is a bit quicker now when the user clicks the stop button on a profile row component
-If a profile fails to populate, the profile manager GUI actually shows that profile as having failed.  Double-clicking it now brings up an exception report dialog instead of the profile viewer window.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@1987 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87