Fixes bug 1337: "Quick expand and collapse for dbtree separator".
Jonathan recommended having a keyboard shortcut but I don't think it's possible. The fix was a one liner call of a JSplitPane method that adds a widget which does the collapse/expand, and it doesn't seem like there is a separate method that we can call to directly collapse/expand the divider.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2001 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87