Fix for bug 1526 http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1526
SQLIndex.getDerivedInstance did not handle indices on expressions properly, throwing an NPE. This fix should allow it to derive such indices. Also updated the documentation for getDerivedInstance based on my understanding of the method, and added a simple regression test that would fail with an NPE with the old code.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2045 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87