Big refactoring of the FileValidator interface and implementations.  It is no longer just for files, so it's called UserPrompter now.
This refactoring doesn't have any visible effect on the existing user interface, but the API is now significantly different. It's actually simpler, yet much more flexible.

The main reason for this change is to pave the way for having other parts of the Architect's business logic be able to ask the user to make decisions in a Swing environment, but still be able to press on without user input in a headless environment.  The specific immediate need is twofold: We want to prompt the user about orphaned profiles when a source database connection is being removed, and we also want to be able to prompt the user about breaking ETL lineage when removing a source connection that has been reverse engineered into the playpen. Those changes aren't in this commit, but will be coming soon to a repository near you.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2046 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87