Added a new behaviour to SQLObject: Now a new type of event is fired before every attempt to remove one or more children of a SQLObject. This new event is a SQLObjectPreEvent, and it gives listeners the chance to veto the real event from actually taking place. The application is to allow various parts of the system to ensure ancestors of SQLObjects they're referencing are not removed, or if they are removed, they have the chance to clean up.
Note that child removal is the only type of pre-event implemented at the moment. We can add the other types (adding children and modifying objects) when and if there is a need.

Coupled with the new UI-neutral UserPrompter system (committed previously), this allows modules like the PlayPen and the ProfileManager to notice when databases are about to be removed, then ask the user if they want to cancel the operation or sever ties to objects under those database connections that are being removed. All in a loosely-coupled, UI-neutral way. Yay.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2051 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87