A partial fix for bug 1447. The problem was when importing the relationships from the source table, it would go through both the imported and exported keys. A self-referencing relationship would be in both, meaning it would get added twice.
Also fixed a bug where importing a self-referencing table twice would result in the second import having it's relationships pointing to the first import of the table.

There is still the issue of all relationships being reverse engineered as identifying, which is also related to bug 1389.

I've raised this issue in the Developer's Mailing list:
http://groups.google.com/group/architect-developers/browse_thread/thread/d70e3e3ee3353f1

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2071 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87