Now the index editor responds to column additions and removals in the table by doing the obvious thing (adding or removing the corresponding table row)
In doing this change, we decided it would be best to migrate away from the DefaultTableModel, and do something a bit more custom that uses actual SQLIndex.Column objects to represent the rows. A lot of error-prone glue code could be eliminated.

Also reformatted a fair bit of code that wasn't directly related to this change.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2096 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87