Fixed bug 1477. Implements synchronization between selection on the dbTree and the playpen.
Removed select and show for now, which will be re-implemented next.
Deletion on dbTree throws an exception occasionally but will be gone when bug 1193 is fixed.

There's minor "bug" where ctrl select does not deselect a relationship on the dbTree. Selecting other objects deselects it as appropriate.
TODO: reimplement select and show in pp.


git-svn-id: http://power-architect.googlecode.com/svn/trunk@2170 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87