Added DDL warning checks that apply to all platforms and ensure that the SET NULL update/delete actions apply to nullable columns, and the SET DEFAULT update/delete actions apply to columns that have a default value.
There is a problem with the GUI where the DDLWarning for the default value always tries to modify the name of the column instead of the specified property (in this case, defaultValue). We will fix that later because it's a pre-existing bug in the GUI and not something we caused in this commit.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2395 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87