Now the OLAPTreeModel works properly with the new MondrianModel classes. It notices when you add and remove children, and fires the appropriate event.
The idea of creating a unified list of children is now encapsulated in a getChildren() method implemented in every OLAPObject. Formerly, it was inside the TreeModel because we had limited control over the code generation before we started doing it ourselves from scratch.

The way the unified child lists are created is currently crude and wasteful, but we're holding off on optimization until it's proven slow in the real world (which will probably never happen).

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2457 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87