Reading and writing MondrianXMLs now works! (mostly).  foodMart.xml can successfully be read and be rewrote, though foodMart.xml does not test every element possible.
Some things (like CData) still need to be escaped before this is completed.  There's also some not so pretty code in MondrianXMLWriter.java that could probably be improved (foolishWrite() inparticular).  I believe the reading and writing can be used at this point with little to no problems so I am commiting it. =)

Also included in the build.xml are three new actions to generate the MondrianModel, MondrianXMLReader and MondrianXMLWriter.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2478 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87