The Architect can now load the OLAP schemas that it saves!
A few classes were changed to adapt to support a list of schemas rather than just one.  We also removed the main method in OLAPTreeModel because it is not needed anymore, instead you can use the Architect ImportSchemaAction =). 

Also a new UncloseableInputStream class was added to prevent parsing methods from closing the InputStream when we did not want to close it! 

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2489 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87