Added ability to choose which table a cube uses (later, you'll also be able to choose a view or inline table).
This required a new utility method for finding the ancestor OLAPSession from a given OLAPObject (since that's where the database comes from). We also took this opportunity to remove an old utility method that's no longer needed.

This new SQLObject interaction means the action has to handle an ArchitectException. So we've put an exception dialog with a user-friendly hint at the probable cause in the edit panel initializing code.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2514 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87