OLAPEditSessions are not associated with OLAPSession instead of the child schema.
Moved the OLAP Schema Editor dialog generation code into the OLAPEditSession. Consequently, the same dialog always pops up for a OLAPSession/Schema.

ArchitectSwingSession now keeps a list of the active OLAPEditSessions. It should now be the only place where OLAPEditSessions are created. The edit session also listens to events for when it should dispose itself and the dialog.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2537 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87