Pulled up dragging and dropping support from TablePane to ContainerPane, in preparation for supporting DnD of OLAP pane items.
We found and fixed some code in PlayPen that was trying to produce a DnD image from a JLabel. It works now, and looks great!

On a sad note, we found a lot of broken stuff with relationship handling while testing our changes. We verified that it was all broken before we started, and that our changes have nothing to do with those bugs.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2598 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87