Fixed a two bugs with synchronization between OLAPTree and playPen.
Previously when you selected a level or hierarchy that was in a dimension, which was in a cube, it would try to select it on the playpen, but it does not exist on the playpen.

Also edited synchro such that if you select an item on the tree that is not represented in the playpen, then it will deselect all items in the playpen.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2627 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87