Added a stop button to the SQL Query tool.
Also added a bean class to hold the connection with a statement currently running on the connection. This way we can send the statement a cancel command.

The bean also holds a boolean to tell if the connection has been modified while not in auto commit to better handle when to pop up a window asking to commit or roll back.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2744 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87