Got rid of a hideous secret in the DBTreeModel: it was actually gluing the profile manager to the SQLObjects it had profiles for!
We solved this problem by introducing "client properties" to SQLObject, which serve much the same purpose as client properties of JComponents. Now anything can stash any number of named properties into a SQLObject, and get them back later. Modifying a client property causes a SQLObjectChangedEvent, which was the whole reason for the hideous code in DBTreeModel in the first place. Client properties are not saved with the project, so they are only useful for storing transient or redundant information.

For our initial use of client properties, we now have the ProfileManager storing a count of the number of profiles associated with each SQLTable. This allows the DBTree to notice a normal property change on a tree node and repaint it in the usual way. The DBTreeModel now has no knowledge of the profile manager, which is a good thing.


git-svn-id: http://power-architect.googlecode.com/svn/trunk@2785 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87