Fix for bugs 1334 and 1661: the compare DM panel now explains why the start button is disabled, and it does allow you to create an upgrade script even when the older item is not a physical database.
This also includes a related prerequisite fix: I removed a useless debug statement  in the SQL Script Dialog that was causing an NPE when trying to show the script dialog with a null target database. The SQL Script Dialog was designed to allow this usage in the first place, so now that it's working again, I added doc comments to make it explicit.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2813 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87