Shift-click now selects a range of columns in a table on the playpen.
Changed the way deselection works when multiple columns were selected. 
Previous behaviour: When a range of columns were selected and a column in this range is selected again, it gets deselected while the other columns are still selected.
Now: Behaves as before only when the control key is down as well.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2967 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87