1.Now allows the user to touch the physical name of SQLColumn and SQLTable. Physical name will not be generated or corrected when forward engineer if user has set one but will pop up a warning if it's not valid for the specific database. So far the user has to set the logical name for a column or table, we can generate the physical name for them like before if physical name is blank.
2.Also modified SQLRelationship and SQLIndex to use physical name if it's not empty instead of logical name anymore.

3.Added some test cases for physical name and changed some test cases for SQLColumnSQLRelationship and TableEditPanel.

This is a fix for bug 1784. Notice that the user will lose their logical names if they forward and then reverse engineer. However, they are supposed to save the project to save their logical names. 
Michelle is working on reporting formats, which is the only thing left for this bug fix.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2975 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87