Fix for "relationship lines appear to vanish when selected." (This issue was reported on the forum by bronius).
Now we use SQL Power Orange for selected relations that can't be darkened. We also changed the column highlight to match, so it's now more obvious what the column highlight means in this context.

With additional suggestions from Giulio, we now also double the line thickness when a relationship is selected, since colour alone can be ambiguous.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@2976 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87