CompareDM now refreshes any physical databases before the compare, to fix a bug where running compareDM, executing the given script, and then immediately running a compareDM after the script is run woudl result in the same SQL script being generated. The problem was that the SQLDatabase cache still had the old database structure before the script was run.
Also modified it so that the progress bar will also be running when the refresh happens as well as when the compare is happening, since the refresh could take some time. I've left it as indeterminate when refreshing since I'm not sure if there's a way to get the progress of a current refresh.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3007 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87