NEW - bug 2708: Multiple problems with SQLRelationship http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2708
There are several changes in this commit that got tied together:
-The SQLRelationship is now more tightly connected to its SQLImportedKey. Previously the imported key could be 
disconnected from its relationship adding complexity when the two ends of the relationship floated apart.
-The ArchitectPropertyChangeUndoableEdit is back in the form of SPObjectPropertyChangeEdit. This is required as magic 
needs to be disabled on undo and redo which the other generic undo event does not handle.
-The SPClassVisitor has been updated to only consider constructors in the class being walked through. Previously it was 
able to consider the constructor of inner classes as valid constructors of the outer class.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3243 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87