NEW - bug 2713: Dragging a column into the primary key of a self referencing table throws IOOBE http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2713
Updated based on a library change. The SQLIndex that is the primary key of a table is now always the same index and the positions of columns is tracked by this index instead of values in the columns that need to be normalized. This reduces the intermediate steps of modifying a table to reduce complexity.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3271 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87