NEW - bug 2697: Remove the SPPersisterHelperFactoryImpl http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2697
Corrected the annotations on the classes extending SPObject. This allows the change to the build.xml to create persister
helpers based on the new annotations. There is currently one test in the regress package for the ArchitectProject but
more are to come soon.
The persister helpers can be created by running the compile target in the build.xml. Currently there are unused imports
which cause errors in Eclipse but this will be fixed soon.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3286 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87