NEW - bug 2755: Do not allow changing the name of one user to a different existing user. http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2755
To pass useful messages from the server there is now a 'friendly' version of a persistence exception. This type of exception
can be thrown on the server to notify the user of a persistence exception and that it needs to roll back but it is not a 
critical error in the server so the user can continue to work. This is currently used to pass a 'username exists' type 
message but more uses will exist in the future. These types of exceptions will come up in friendly dialogs without red
text or a stack trace.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3471 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87