ASSIGNED - bug 2682: Implement Save/Load to XML with User Defined Types http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2682
Made an enhancement in which if the type UUID doesn't exist in the loading session, then it will attempt to replace it with an existing type with the same JDBC type.

Note however that this may not end up picking the 'right' type replacement if there's more than one UserDefinedSQLType with the same JDBC type in a session.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3493 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87