The ColumnEditPanel now allows selection of domains, not just data types. Refactored a Popup creator from Wabit into the library. This popup was used in Wabit to choose cubes in an OLAP query. It has been moved into the library to allow the ColumnEditPanel in Architect to use a popup for choosing domains and data types.
A new TreeCellRenderer has been created to render domains and data types in the tree displayed in the Popup, as well as the Domain Editor Panel. Each cell of the tree is rendered to display user friendly text, icons, and tooltips.

If two or more users are modifying the same column type and one user commits the change, the users who have not committed the change will see that their domain/type tree will be highlighted in red or yellow, just like the other components in the panel do.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3538 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87