Two fixes to profiling related to threading. The first problem is the profile results were being copied previously so changes would be done on the foreground thread. This copying was being done on the background thread but needed to be done on the foreground to prevent the getters from creating an inconsistent copy.
The second change was to also copy the table and columns of the profile being copied. This way the table would not be changed while doing the profiling. This was causing profiles to come out empty as the table would end up populating on a different thread after the profile was completed on the background thread.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3634 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87