Fixed some issues with the MySQL, H2, and SQL Server DDL generators. In the MySQL DDL generator, the check constraint methods have been overwritten to perform no-op since it does not support check constraints.
In the SQL Server DDL generator, column level constraints are generated on the table level since only one named check constraint is allowed on the column level. These table level constraints are tagged with a prefix to identify where the constraint is actually applied on.

Similarly, in H2, it does not allow column level check constraints at all. They must be added on the table level, and tagged with a prefix as well.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3700 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87