Removed the defaultValue property from ConstructorParameter and removed the PRIMITIVE enum from PropertyType within that annotation. Nothing actually uses it.
Originally, a primitive constructor parameter meant it was not a child or property. For example, it could be a boolean flag that indicated whether a certain operation within the constructor should be performed. However, after discussing with Matt and Thomas, we agreed that a better solution would be to just create another constructor that feeds in a default value for the flag and use that one instead.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3743 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87