* Added support for allowing UserDefinedSQLTypeSnapshots and DomainCategorySnapshots to be created from the client.
This is primarily to solve the issue where old revisions of server-based projects needed their types to maintain the state they were in at those revision times, since types are normally contained in the system workspace.

The creation currently occurs in the client, but I would really prefer it to be done on the server, but I haven't been able to figure out a way to do so within the server backend.

* The ArchitectClientSideSession's getters for types and domain categories need to merge the system types and categories with the snapshotted ones.

* SQLTypeTreeModel.compare threw exceptions if two objects with different parents were compared. However, since snapshotted types are typically parented by the project, while types are parented by the system workspace, this can no longer be the case.

* Added a boolean flag in UserDefinedSQLTypeSnapshot to mark whether or not the snapshot is being used to snapshot a domain. This makes it easier to support having separate snapshots of a type for the purpose of being referenced by a domain, and of the type itself.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3784 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87