Fixed a bug introduced by the new user defined types and domain snapshots feature. Adding in new snapshots must update all of the existing user defined type's upstream type references. A new upstream type updating listener handles this case by listening to child added events on ArchitectProject where the child UserDefineSQLTypeSnapshot is not a domain snapshot. It looks through all of the descendant UserDefinedSQLTypes where its upstream type's UUID is equal to the snapshot's original UUID, and updates the upstream type references to the new snapshot type.
The SQLTypeTreeModel has also been changed so that it only retrieves data types and domains directly from the session when the model is first created. Subsequent method calls within the model that define the tree model's structure should only reflect the model at creation time, and not any updates that happen afterwards.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3790 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87