NEW - bug 2917: Exceptions when editing security settings http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2917
Fixed the bug where having multiple Grants within a Group that refer to an ArchitectSwingProject, and opening the GroupEditorPanel would throw an exception. This was caused by the change in definition of what a system level Grant is. A system level grant must have a null subject. The check has been changed so that only 1 system level ArchitectSwingProject Grant is allowed.

Also, added documentation to PrivilegesEditorPanel.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3816 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87