Added support so that type snapshots get created from reverse engineering. This required combining the setting of the upstream type to a system type, and then to a snapshot type to occur in the same transaction, so modifications to the SPObjectSnapshotHierarchyListener were made to accomodate this. Also removed the UpstreamTypeUpdateListener since reverse engineered columns use snapshots right away, so it is no longer necessary.
One issue that remains is that when loading a project, the update listeners appear to be not attached or not working as changes to domains and types will not result in a refresh icon appearing over the used types in the project in the dbtree. However, you can still rightclick and update to retrieve changes.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3848 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87