NEW - bug 2942: Test Connection with wrong password says the test is succesful http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2942
The issue was that if a valid connection has already been established, the cookie store contains the server information. If you modify the password of this connection and test it again, it would look in the cookie store. It has been changed so that the cookie store is cleared every time Test Connection is clicked.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3870 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87