NEW - bug 2943: Max Named Users is not checked http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2943
The number of users is now respected for the license file. If the user tries to create more users than their license allows the transaction is rejected with a friendly error message.

With this fix comes another fix on the error messages. Previously if a friendly error message was sent from the server the NetworkConflictResolver would catch it and display a friendly dialog. However, the persistence layer was relying on an exception being thrown to interrupt the commit and cause a correct roll back which was not being thrown. This caused the commit to continue after some objects rolled back and a commit be sent to persisters that had their transaction count set to 0 from the rollback. Now these exceptions move up the stack and there is a new exception handler that displays nicer dialogs for friendly exceptions that we have defined to be user errors.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3877 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87