NEW - bug 2914: Reverse engineering can bring user out of sync with server http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2914
I believe this should fix the issue with reverse engineering. The NetworkConflictResolver, on flush(), will decode the JSON and notify all of the UpdateListeners that an update is going through. This call needs to occur on a foreground thread because if a component is being dragged in the PlayPen, it needs to be interrupted and rolled back correctly before applying the change from the server.

The UpdateListener for PlayPenComponent now rolls back the drag transaction on a preUpdatePerformed call instead of an updatePerformed call, and removes the FloatingContainerPaneListeners from the PlayPen.

git-svn-id: http://power-architect.googlecode.com/svn/trunk@3887 ef18fcdb-8a34-0410-ae7e-c1cd3e483e87